refactor(scrapers): move ObjectToXmlConverter to scrapers module
Moved ObjectToXmlConverter from lib/html2rss to lib/html2rss/scrapers and updated references. Enhanced logging in call method.
